se cambio tipo de dato int a long

+++ src/main/java/com/compi/elitewings/Controllers/CelebrityController.java
@@ -40,7 +40,7 @@ public class CelebrityController {
     @PutMapping("/celebrities/{id}")
     public String updateCelebrity(@PathVariable("id") long id, @RequestBody Celebrity celebrity){
         try {
-            celebrity.setId((int) id);
+            celebrity.setId(id);
             this.serviceCelebrity.updateCelebrity(celebrity);
             return "Celebrity with ID " + id + " updated successfully.";
         } catch (RuntimeException e) {
